# _extract_sle/Store-Link-Engine/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ziadah Product Intelligence Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List all registered stores
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Register a new store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreBody"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{id}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get a single store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStore
      tags: [stores]
      summary: Update a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreBody"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: Delete a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{id}/sync:
    post:
      operationId: syncStore
      tags: [stores]
      summary: Fetch and sync products from the store URL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "400":
          description: Sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{id}/analyze:
    post:
      operationId: analyzeStore
      tags: [analysis]
      summary: Run AI analysis to identify main product, cross-sells and upsells
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Analysis failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{id}/products:
    get:
      operationId: listStoreProducts
      tags: [products]
      summary: List all products for a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{id}/analysis:
    get:
      operationId: getLatestAnalysis
      tags: [analysis]
      summary: Get the latest analysis results for a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Latest analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: No analysis found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/{id}/recommendations:
    get:
      operationId: getRecommendations
      tags: [analysis]
      summary: Get product recommendations summary for a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommendations summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationsSummary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Store:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        platform:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, synced, analyzing, analyzed, error]
        productCount:
          type: integer
        lastSyncedAt:
          type: ["string", "null"]
        lastAnalyzedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - url
        - platform
        - status
        - productCount
        - lastSyncedAt
        - lastAnalyzedAt
        - createdAt
        - updatedAt
    CreateStoreBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
      required:
        - name
        - url
    UpdateStoreBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
    Product:
      type: object
      properties:
        id:
          type: integer
        storeId:
          type: integer
        externalId:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        compareAtPrice:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        productUrl:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        inStock:
          type: boolean
        role:
          type: ["string", "null"]
          enum: [main, cross_sell, upsell, null]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - storeId
        - externalId
        - title
        - description
        - price
        - compareAtPrice
        - imageUrl
        - productUrl
        - category
        - tags
        - inStock
        - role
        - createdAt
        - updatedAt
    SyncResult:
      type: object
      properties:
        synced:
          type: integer
        total:
          type: integer
        platform:
          type: string
        message:
          type: string
      required:
        - synced
        - total
        - platform
        - message
    ProductRecommendation:
      type: object
      properties:
        productId:
          type: integer
        title:
          type: string
        imageUrl:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        reason:
          type: string
      required:
        - productId
        - title
        - imageUrl
        - price
        - reason
    AnalysisResult:
      type: object
      properties:
        storeId:
          type: integer
        analyzedAt:
          type: string
        mainProduct:
          $ref: "#/components/schemas/ProductRecommendation"
        crossSells:
          type: array
          items:
            $ref: "#/components/schemas/ProductRecommendation"
        upsells:
          type: array
          items:
            $ref: "#/components/schemas/ProductRecommendation"
        summary:
          type: string
      required:
        - storeId
        - analyzedAt
        - mainProduct
        - crossSells
        - upsells
        - summary
    RecommendationsSummary:
      type: object
      properties:
        storeId:
          type: integer
        storeName:
          type: string
        totalProducts:
          type: integer
        mainProductTitle:
          type: ["string", "null"]
        crossSellCount:
          type: integer
        upsellCount:
          type: integer
        lastAnalyzedAt:
          type: ["string", "null"]
        status:
          type: string
      required:
        - storeId
        - storeName
        - totalProducts
        - mainProductTitle
        - crossSellCount
        - upsellCount
        - lastAnalyzedAt
        - status