# lib/api-zod/src/generated/types/productRecommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ziadah Product Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductRecommendation {
  productId: number;
  title: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  price: number | null;
  reason: string;
}